(BSR) feat(user): move useGetDepositAmountByAge in shared/user
