+++ src/features/profile/components/Header/NonBeneficiaryHeader/NonBeneficiaryHeader.tsx
@@ -9,8 +9,8 @@ import { IdentityCheckPendingBadge } from 'features/profile/components/Badges/Id
 import { SubscriptionMessageBadge } from 'features/profile/components/Badges/SubscriptionMessageBadge'
 import { YoungerBadge } from 'features/profile/components/Badges/YoungerBadge'
 import { Subtitle } from 'features/profile/components/Subtitle/Subtitle'
-import { useGetDepositAmountsByAge } from 'features/user/helpers/useGetDepositAmountsByAge/useGetDepositAmountsByAge'
 import { formatToSlashedFrenchDate } from 'libs/dates'
+import { useGetDepositAmountsByAge } from 'shared/user/useGetDepositAmountsByAge'
 import { PageHeader } from 'ui/components/headers/PageHeader'
 import { BannerWithBackground } from 'ui/components/ModuleBanner/BannerWithBackground'
 import { BicolorUnlock } from 'ui/svg/icons/BicolorUnlock'
